+++ src/modules/ingredient/ingredient.types.ts
@@ -1,5 +1,3 @@
-import { IngredientUnit } from '../../common/enums';
-
 export interface IngredientDataset {
     category: IngredientCategory;
     data: IngredientType[];
@@ -9,7 +7,7 @@ export interface MealIngredient {
     amount: number;
     imageUrl: string;
     name: IngredientType;
-    unit: IngredientUnit;
+    unit: string;
 }
 
 export type IngredientCategory = 'breads' | 'dairy-and-eggs' | 'fish-and-seafood' | 'fruits' | 'meats' | 'oils-and-fats' | 'pasta' | 'seeds-and-nuts' | 'spices' | 'vegetables';